Auomatic backend dispatch for samplers
