+++ Examples/MPS/mps.py
@@ -1,11 +1,10 @@
 import netket as nk
 import numpy as np
 import jax
-from jax.experimental.optimizers import sgd as JaxSgd
-import cProfile
+from jax.experimental.optimizers import adam as Adam
 
 # 1D Lattice
-L = 16
+L = 20
 g = nk.graph.Hypercube(length=L, n_dim=1, pbc=True)
 
 # Hilbert space of spins on the graph
@@ -17,22 +16,16 @@ ma = nk.machine.MPSPeriodic(hi, bond_dim=4, diag=False, symperiod=None, dtype=co
 ma.jax_init_parameters(seed=1232)
 
 # Jax Sampler
-sa = nk.sampler.jax.MetropolisLocal(machine=ma, n_chains=8)
+sa = nk.sampler.MetropolisLocal(machine=ma, n_chains=2)
 
 # Using a Jax Optimizer
-j_op = JaxSgd(0.01)
-op = nk.optimizer.Jax(ma, j_op)
+op = nk.optimizer.Jax(ma, Adam(0.01))
 
-
-# Stochastic Reconfiguration
-sr = nk.optimizer.SR(diag_shift=0.1)
 
 # Create the optimization driver
-gs = nk.Vmc(
-    hamiltonian=ha, sampler=sa, optimizer=op, n_samples=1000, sr=None, n_discard=0
-)
+gs = nk.Vmc(hamiltonian=ha, sampler=sa, optimizer=op, n_samples=1000, sr=None)
 
 # The first iteration is slower because of start-up jit times
 gs.run(out="test", n_iter=1)
 
-gs.run(out="test", n_iter=10000)
+gs.run(out="test", n_iter=300)
