+++ netket/machine/__init__.py
@@ -13,8 +13,4 @@ if torch_available:
     from .torch import Torch, TorchLogCosh, TorchView
 
 
-# def MPSPeriodicDiagonal(hilbert, bond_dim, symperiod=-1):
-#     return MPSPeriodic(hilbert, bond_dim, diag=True, symperiod=symperiod)
-
-
 from . import density_matrix
