+++ netket/sampler/__init__.py
@@ -1,17 +1,10 @@
 from .abstract_sampler import AbstractSampler
 
-from . import numpy
-from ..utils import jax_available
+from .metropolis_hastings import *
 
-if jax_available:
-    from . import jax
+from .metropolis_local import *
+from .metropolis_exchange import *
+from .metropolis_hamiltonian import *
+from .custom_sampler import *
 
-from .metropolis_hastings import MetropolisHastings, MetropolisHastingsPt
-
-# from .metropolis_hastings_pt import MetropolisHastingsPt
-# from .metropolis_local import MetropolisLocal, MetropolisLocalPt
-from .metropolis_exchange import MetropolisExchange, MetropolisExchangePt
-
-# from .metropolis_hamiltonian import MetropolisHamiltonian, MetropolisHamiltonianPt
-# from .custom_sampler import CustomSampler, CustomSamplerPt
-# from .exact_sampler import ExactSampler
+from .exact_sampler import *
