+++ netket/sampler/numpy/__init__.py
@@ -1,8 +1,7 @@
 from .metropolis_hastings import MetropolisHastings
-from .metropolis_exchange import ExchangeKernel as NumpyExchangeKernel
-from functools import singledispatch
+from .metropolis_hastings_pt import MetropolisHastingsPt
 
-
-@singledispatch
-def ExchangeKernel(machine, dmax):
-    return NumpyExchangeKernel(machine.hilbert, dmax)
+from .local_kernel import _LocalKernel
+from .exchange_kernel import _ExchangeKernel
+from .custom_kernel import _CustomKernel
+from .hamiltonian_kernel import _HamiltonianKernel
